correct images and new common component
