+++ src/components/common/ProductBox/ProductBox.js
@@ -3,52 +3,17 @@ import PropTypes from 'prop-types';
 
 import styles from './ProductBox.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faExchangeAlt, faShoppingBasket } from '@fortawesome/free-solid-svg-icons';
-import { faStar as faHeart } from '@fortawesome/free-regular-svg-icons';
+import { faShoppingBasket } from '@fortawesome/free-solid-svg-icons';
 import Button from '../Button/Button';
-import { useDispatch, useSelector } from 'react-redux';
-import {
-  getComparedProducts,
-  toggleCompare,
-  toggleFavorite,
-  updateUserRating,
-} from '../../../redux/productsRedux';
-import StarRating from '../../features/StarRating/StarRating';
+import { useSelector } from 'react-redux';
+import { getOne } from '../../../redux/productsRedux';
+import ProductBoxContent from './ProductBoxContent';
 import { useTranslation } from 'react-i18next';
 
-const ProductBox = ({
-  id,
-  name,
-  price,
-  promo,
-  stars,
-  oldPrice,
-  favorite,
-  compare,
-  userRating: initialUserRating,
-}) => {
+const ProductBox = ({ name, price, promo, stars, oldPrice, ...props }) => {
   const { t } = useTranslation();
-  const [userRating, setUserRating] = useState(initialUserRating || 0);
 
-  const handleRatingChange = (id, rating) => {
-    setUserRating(rating);
-    dispatch(updateUserRating(id, rating));
-  };
-
-  const dispatch = useDispatch();
-  const products = useSelector(getComparedProducts);
-
-  const handleToggleFavoriteProduct = e => {
-    e.preventDefault();
-    dispatch(toggleFavorite(id));
-  };
-
-  const handleToggleCompareProduct = e => {
-    e.preventDefault();
-    if (products.length < 4 || compare === true) {
-      dispatch(toggleCompare(id));
-    }
-  };
+  const product = useSelector(state => getOne(state, props.id));
 
   return (
     <div className={styles.root}>
@@ -56,7 +21,7 @@ const ProductBox = ({
         <img
           className={styles.image}
           alt={name}
-          src={`${process.env.PUBLIC_URL}/images/products/${id}.jpg`}
+          src={`${process.env.PUBLIC_URL}/images/products/${props.id}.jpg`}
         />
         {price < oldPrice && <div className={styles.sale}>{promo}</div>}
         <div className={styles.buttons}>
@@ -67,43 +32,7 @@ const ProductBox = ({
           </Button>
         </div>
       </div>
-      <div className={styles.content}>
-        <h5>{name}</h5>
-        <div className={styles.stars}>
-          <StarRating
-            defaultRating={stars}
-            clientRating={userRating}
-            onRatingChange={rating => handleRatingChange(id, rating)}
-          />
-        </div>
-      </div>
-      <div className={styles.line}></div>
-      <div className={styles.actions}>
-        <div className={styles.outlines}>
-          <Button
-            className={favorite && styles.favorite}
-            onClick={handleToggleFavoriteProduct}
-            variant='outline'
-          >
-            <FontAwesomeIcon icon={faHeart}>{t('label.favorite')}</FontAwesomeIcon>
-          </Button>
-          <Button
-            className={compare && styles.compare}
-            onClick={handleToggleCompareProduct}
-            variant='outline'
-          >
-            <FontAwesomeIcon icon={faExchangeAlt}>
-              {t('label.favorite')}
-            </FontAwesomeIcon>
-          </Button>
-        </div>
-        <div className={styles.price}>
-          {oldPrice && <p>${oldPrice}</p>}
-          <Button noHover variant='small'>
-            $ {price}
-          </Button>
-        </div>
-      </div>
+      <ProductBoxContent key={product.id} {...product} />
     </div>
   );
 };
@@ -115,10 +44,8 @@ ProductBox.propTypes = {
   price: PropTypes.number,
   promo: PropTypes.string,
   stars: PropTypes.number,
-  favorite: PropTypes.bool,
-  compare: PropTypes.bool,
   oldPrice: PropTypes.number,
-  userRating: PropTypes.number,
+  id: PropTypes.string,
 };
 
 export default ProductBox;
