+++ src/components/common/ProductBox/ProductBox.module.scss
@@ -2,8 +2,6 @@
 
 .root {
   background-color: $pb-box-bg;
-  border: 1px solid $pb-box-border;
-  margin-bottom: 2rem;
 
   .photo {
     position: relative;
@@ -43,81 +41,6 @@
       width: 100%;
       bottom: 0;
     }
-  }
-
-  .content {
-    text-align: center;
-    padding: 20px;
-
-    h5 {
-      color: $primary;
-      font-size: 16px;
-      line-height: 24px;
-      letter-spacing: 1px;
-    }
-
-    .stars {
-      a {
-        text-decoration: none;
-        color: $pb-stars-border;
-      }
-    }
-  }
-
-  .line {
-    margin: 0 10px;
-    border-bottom: 1px solid $pb-box-line;
-    position: relative;
-
-    &::before {
-      content: "";
-      position: absolute;
-      bottom: -1px;
-      left: 50%;
-      width: 30px;
-      border-bottom: 4px solid $pb-line-ornament;
-      transform: translateX(-50%);
-    }
-  }
-
-  .actions {
-    padding: 15px 10px 10px 10px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .outlines {
-    .favorite {
-    background-color: $primary;
-    color: $light-text-color;
-    }
-
-    .compare {
-      background-color: $primary;
-      color: $light-text-color;
-    }
-  }
-
-  &:hover {
-    .buttons {
-      visibility: visible;
-    }
-
-    .price > div {
-      background-color: $primary;
-    }
-  }
-}
-
-.price {
-  display: flex;
-
-  p {
-    margin: auto 3px;
-    color: $color-gray;
-    text-decoration: line-through;
-    font-weight: 200;
   }
 }
