+++ src/components/common/ProductBox/ProductBoxContent.module.scss
@@ -1,0 +1,107 @@
+@import "../../../styles/settings.scss";
+
+.root {
+  background-color: #ffffff;
+  border: 1px solid #e1e1e1;
+  margin-bottom: 2rem;
+
+  .content {
+    text-align: center;
+    padding: 20px;
+
+    h5 {
+      color: $primary;
+      font-size: 16px;
+      line-height: 24px;
+      letter-spacing: 1px;
+      white-space: normal;
+    }
+
+    .stars {
+      a {
+        text-decoration: none;
+        color: rgb(42, 42, 42);
+      }
+    }
+  }
+
+  .line {
+    margin: 0 10px;
+    border-bottom: 1px solid #2a2a2a;
+    position: relative;
+
+    &::before {
+      content: "";
+      position: absolute;
+      bottom: -1px;
+      left: 50%;
+      width: 30px;
+      border-bottom: 4px solid $primary;
+      transform: translateX(-50%);
+    }
+  }
+
+  .actions {
+    padding: 15px 10px 10px 10px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+
+  .outlines {
+    .favorite {
+      color: $danger;
+
+      & > svg {
+        scale: 1.2;
+      }
+    }
+
+    .compare {
+      color: $primary;
+
+      & > svg {
+        scale: 1.2;
+      }
+    }
+    .viewHidden {
+      visibility: hidden;
+    }
+  }
+
+  &:hover {
+    .buttons {
+      visibility: visible;
+    }
+
+    .price > div {
+      background-color: $primary;
+    }
+  }
+}
+
+.price {
+  display: flex;
+
+  p {
+    margin: auto 3px;
+    color: $color-gray;
+    text-decoration: line-through;
+    font-weight: 200;
+  }
+}
+
+@media (max-width: 577px) {
+  .root {
+    margin: 20px auto;
+    max-width: 300px;
+  }
+}
+
+@media (min-width: 578px) and (max-width: 991px) {
+  .root {
+    margin: 20px auto;
+    max-width: 300px;
+  }
+}
