+++ src/components/features/Promoted/Promoted.module.scss
@@ -281,18 +281,3 @@
   }
 }
 
-/*
-@media (max-width: 577px) {
-  .root {
-    margin: 20px auto;
-    max-width: 300px;
-  }
-}
-
-@media (min-width: 578px) and (max-width: 991px) {
-  .root {
-    margin: 20px auto;
-    max-width: 300px;
-  }
-}
- */
